added user input for area to simulate for. people scale to fit in the window.

+++ src/Window.java
@@ -5,9 +5,10 @@ public class Window extends JFrame
 {
     public Window()
     {
+
         Simulation sim = new Simulation();
         
-        getContentPane().setPreferredSize(new Dimension(sim.getX_BOUND_MAX()+sim.getPerson(0).getRadius(), sim.getY_BOUND_MAX()+sim.getPerson(0).getRadius()));
+        getContentPane().setPreferredSize(new Dimension((int)(sim.getX_BOUND_MAX()+sim.getPerson(0).getRadius()), (int) (sim.getY_BOUND_MAX()+sim.getPerson(0).getRadius())));
         getContentPane().setMinimumSize(new Dimension(640, 480));
         getContentPane().setMaximumSize(new Dimension(1920, 1080));
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
